backend/requirements.txt
Flask==2.3.3
prefect==2.14.18
python-dotenv==1.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9   
mysql-connector-python==8.3.0 
Werkzeug==3.0.1
marshmallow==3.21.1
flask-marshmallow==0.15.0
gunicorn==22.0.0  
bcrypt==4.1.2              
PyJWT==2.8.0              
python-multipart==0.0.7      
cloudinary
pillow>=10.2.0  
griffe==0.37.0
pandas==2.2.2